+++ tests/pyscript/foo3.py
@@ -1,0 +1,1 @@
+app.foo('11', '22', '33', '44', '55', '66', counter=3, trueval=False, constval=False)