+++ tests/pyscript/media_movies_add2.py
@@ -1,0 +1,1 @@
+app.media.movies.add('My Movie', 'PG-13', actor=('Mark Hamill'), director=('George Lucas', 'J. J. Abrams'))