+++ CHANGELOG.md
@@ -7,6 +7,7 @@
         * See the [tab_autocompletion.py](https://github.com/python-cmd2/cmd2/blob/master/examples/tab_autocompletion.py) example for a demonstration of how to use this feature
     * ``cmd2`` no longer depends on the ``six`` module
     * ``cmd2`` is now a multi-file Python package instead of a single-file module
+    * New pyscript approach that provides a pythonic interface to commands in the cmd2 application.
     * Switch command parsing from pyparsing to custom code which utilizes shlex.
         * The object passed to do_* methods has changed. It no longer is the pyparsing object, it's a new Statement object, which is a subclass of ``str``. The statement object has many attributes which give you access to various components of the parsed input. If you were using anything but the string in your do_* methods, this change will require you to update your code.
         * ``commentGrammers`` is no longer supported or available. Comments are C-style or python style.
@@ -26,6 +27,7 @@
         * Alternatively, see the [argparse_example.py](https://github.com/python-cmd2/cmd2/blob/master/examples/argparse_example.py)
     * Deleted ``cmd_with_subs_completer``, ``get_subcommands``, and ``get_subcommand_completer``
         * Replaced by default AutoCompleter implementation for all commands using argparse
+    * Deleted support for old method of calling application commands with ``cmd()`` and ``self``
 * Python 2 no longer supported
     * ``cmd2`` now supports Python 3.4+